# server/docs/swagger.yaml
definitions:
  serviceAuth.tLoginRes:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: integer
      token:
        type: string
      username:
        type: string
    type: object
  serviceAuth.tVerifyEmailRes:
    properties:
      email:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /check-auth:
    post:
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: bool
      summary: Check Auth
      tags:
      - Auth
  /forgot-password:
    post:
      parameters:
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      responses:
        "200":
          description: go to check your email
          schema:
            type: string
      summary: Send verification code to email
      tags:
      - Auth
  /login:
    post:
      parameters:
      - description: Username
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: Password
        in: body
        name: password
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serviceAuth.tLoginRes'
      summary: Login
      tags:
      - Auth
  /logout:
    post:
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Logout
      tags:
      - Auth
  /reset-password:
    post:
      parameters:
      - description: Toke
        in: path
        name: token
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      - description: Confirmation
        in: formData
        name: confirmation
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: bool
      summary: ResetPassword
      tags:
      - Auth
  /sign-up:
    post:
      parameters:
      - description: UsreId
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: AddContacts
      tags:
      - Contacts
  /user/page:
    get:
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: PageNo
        in: query
        name: page_no
        required: true
        type: integer
      - description: PageSize
        in: query
        name: page_size
        required: true
        type: integer
      - description: Username
        in: query
        name: username
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: UserPage
      tags:
      - User
  /verify-email:
    post:
      parameters:
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Code
        in: formData
        name: code
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serviceAuth.tVerifyEmailRes'
      summary: Submit verification code
      tags:
      - Auth
swagger: "2.0"